+++ pkg/models/redis-json.go
@@ -1,0 +1,12 @@
+package models
+
+/**
+ * RedisJSON Commands
+ */
+const (
+	JsonType    = "json.type"
+	JsonGet     = "json.get"
+	JsonObjKeys = "json.objkeys"
+	JsonObjLen  = "json.objlen"
+	JsonArrLen  = "json.arrlen"
+)